// src/app/api/nakup/model/paymentOrderDTO.ts
/**
 * Nakup
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ElementDTO } from './elementDTO';


export interface PaymentOrderDTO { 
    id_buyer?: number;
    id_seller?: number;
    amount?: number;
    currency?: string;
    id_order?: string;
    redirect_url?: string;
    return_url?: string;
    cancel_url?: string;
    street?: string;
    house_number?: string;
    city?: string;
    postal_code?: string;
    country?: string;
    items?: Array<ElementDTO>;
}